requirements.txt
# Project Dependencies for Document Retrieval Agent
# Install with: pip install -r requirements.txt
# Or use: uv sync (recommended)

# Core ADK & Agent Libraries
google-cloud-aiplatform[adk,agent-engines]>=1.108.0
google-adk>=1.10.0
google-genai>=0.1.0

# Google Cloud
google-auth>=2.25.0
google-cloud-storage>=2.10.0

# Configuration & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (Optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

